deduplicate util scripts
