--- utils/format-all.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-set -eu
-
-cd $(dirname $0)/..
-for n in */format.sh ; do
-    $n
-done
